refactor(api): streamline canvas template handling and update API endpoints
- Refactored CanvasTemplate model in Prisma schema to include a new shareId field and adjusted related properties.
- Removed deprecated canvas template methods from CanvasController and CanvasService.
- Updated API endpoints in OpenAPI schema and service generation to reflect new template routes.
- Adjusted DTOs to align with the new structure, ensuring consistency across the application.
