+++ apps/api/src/canvas/canvas.controller.ts
@@ -11,11 +11,7 @@ import {
 import { JwtAuthGuard } from '@/auth/guard/jwt-auth.guard';
 import { CanvasService } from './canvas.service';
 import { LoginedUser } from '@/utils/decorators/user.decorator';
-import {
-  canvasPO2DTO,
-  canvasTemplateCategoryPO2DTO,
-  canvasTemplatePO2DTO,
-} from '@/canvas/canvas.dto';
+import { canvasPO2DTO } from '@/canvas/canvas.dto';
 import { buildSuccessResponse } from '@/utils';
 import {
   User,
@@ -24,8 +20,6 @@ import {
   AutoNameCanvasRequest,
   AutoNameCanvasResponse,
   DuplicateCanvasRequest,
-  CreateCanvasTemplateRequest,
-  UpdateCanvasTemplateRequest,
 } from '@refly-packages/openapi-schema';
 
 @Controller('v1/canvas')
@@ -93,44 +87,5 @@ export class CanvasController {
   ): Promise<AutoNameCanvasResponse> {
     const data = await this.canvasService.autoNameCanvas(user, body);
     return buildSuccessResponse(data);
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('template/list')
-  async listCanvasTemplates(
-    @LoginedUser() user: User,
-    @Query('categoryId') categoryId: string,
-    @Query('scope', new DefaultValuePipe('public')) scope: 'public' | 'private',
-    @Query('page', new DefaultValuePipe(1), ParseIntPipe) page: number,
-    @Query('pageSize', new DefaultValuePipe(10), ParseIntPipe) pageSize: number,
-  ) {
-    const templates = await this.canvasService.listCanvasTemplates(user, {
-      page,
-      pageSize,
-      scope,
-      categoryId,
-    });
-    return buildSuccessResponse(templates.map(canvasTemplatePO2DTO));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post('template/create')
-  async createCanvasTemplate(@LoginedUser() user: User, @Body() body: CreateCanvasTemplateRequest) {
-    const template = await this.canvasService.createCanvasTemplate(user, body);
-    return buildSuccessResponse(canvasTemplatePO2DTO(template));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Post('template/update')
-  async updateCanvasTemplate(@LoginedUser() user: User, @Body() body: UpdateCanvasTemplateRequest) {
-    const template = await this.canvasService.updateCanvasTemplate(user, body);
-    return buildSuccessResponse(canvasTemplatePO2DTO(template));
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('template/category/list')
-  async listCanvasTemplateCategories() {
-    const categories = await this.canvasService.listCanvasTemplateCategories();
-    return buildSuccessResponse(categories.map(canvasTemplateCategoryPO2DTO));
   }
 }
