+++ apps/api/src/canvas/canvas.dto.ts
@@ -1,14 +1,5 @@
-import {
-  Canvas as CanvasModel,
-  CanvasTemplateCategory as CanvasTemplateCategoryModel,
-  CanvasTemplate as CanvasTemplateModel,
-} from '@prisma/client';
-import {
-  Canvas,
-  CanvasTemplate,
-  CanvasTemplateCategory,
-  Entity,
-} from '@refly-packages/openapi-schema';
+import { Canvas as CanvasModel } from '@prisma/client';
+import { Canvas, Entity } from '@refly-packages/openapi-schema';
 import { pick } from '@/utils';
 
 export interface SyncCanvasEntityJobData {
@@ -37,37 +28,5 @@ export function canvasPO2DTO(canvas: CanvasModel & { minimapUrl?: string }): Can
     ...pick(canvas, ['canvasId', 'title', 'minimapUrl', 'minimapStorageKey']),
     createdAt: canvas.createdAt.toJSON(),
     updatedAt: canvas.updatedAt.toJSON(),
-  };
-}
-
-export function canvasTemplatePO2DTO(
-  template: CanvasTemplateModel & { category?: CanvasTemplateCategoryModel },
-): CanvasTemplate {
-  return {
-    ...pick(template, [
-      'title',
-      'uid',
-      'version',
-      'templateId',
-      'categoryId',
-      'originCanvasId',
-      'shareUser',
-      'description',
-      'language',
-    ]),
-    createdAt: template.createdAt.toJSON(),
-    updatedAt: template.updatedAt.toJSON(),
-    shareUser: JSON.parse(template.shareUser || '{}'),
-    category: template.category ? canvasTemplateCategoryPO2DTO(template.category) : undefined,
-  };
-}
-
-export function canvasTemplateCategoryPO2DTO(
-  category: CanvasTemplateCategoryModel,
-): CanvasTemplateCategory {
-  return {
-    ...pick(category, ['categoryId', 'name', 'labelDict', 'descriptionDict']),
-    labelDict: JSON.parse(category.labelDict || '{}'),
-    descriptionDict: JSON.parse(category.descriptionDict || '{}'),
   };
 }
