+++ apps/api/src/template/template.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { TemplateController } from './template.controller';
+
+describe('TemplateController', () => {
+  let controller: TemplateController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [TemplateController],
+    }).compile();
+
+    controller = module.get<TemplateController>(TemplateController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
