+++ apps/api/src/template/template.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { TemplateService } from './template.service';
+
+describe('TemplateService', () => {
+  let service: TemplateService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [TemplateService],
+    }).compile();
+
+    service = module.get<TemplateService>(TemplateService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
